+++ Test/WinGet.Unit.Tests.ps1
@@ -10,6 +10,7 @@ Describe 'basic package search operations' {
 	Context 'without additional arguments' {
 		BeforeAll {
 			$package = 'CPUID.HWMonitor'
+			$version = '1.44'
 		}
 
 		It 'gets a list of latest installed packages' {
@@ -17,6 +18,9 @@ Describe 'basic package search operations' {
 		}
 		It 'searches for the latest version of a package' {
 			Find-Package -Provider $WinGet -Name $package | Where-Object {$_.Name -contains $package} | Should -Not -BeNullOrEmpty
+		}
+		It 'returns all available versions of a package' {
+			Find-Package -Provider $WinGet -Name $package -AllVersions | Where-Object {$_.Version -eq $version} | Should -Not -BeNullOrEmpty
 		}
 		It 'returns additional package metadata' {
 			Find-Package -Provider $WinGet -Name $package | Select-Object -ExpandProperty FullPath | Should -Not -BeNullOrEmpty
@@ -31,19 +35,20 @@ Describe 'DSC-compliant package installation and uninstallation' {
 	Context 'without additional arguments' {
 		BeforeAll {
 			$package = 'CPUID.HWMonitor'
+			$version = '1.44'
 		}
 
-		It 'searches for the latest version of a package' {
-			Find-Package -Provider $WinGet -Name $package | Where-Object {$_.Name -contains $package} | Should -Not -BeNullOrEmpty
+		It 'searches for a specific version of a package' {
+			Find-Package -Provider $WinGet -Name $package -RequiredVersion $version | Where-Object {$_.Name -contains $package} | Should -Not -BeNullOrEmpty
 		}
-		It 'silently installs the latest version of a package' {
-			Install-Package -Provider $WinGet -Name $package -Force | Where-Object {$_.Name -contains $package} | Should -Not -BeNullOrEmpty
+		It 'silently installs a specific version of a package' {
+			Install-Package -Provider $WinGet -Name $package -RequiredVersion $version -Force | Where-Object {$_.Name -contains $package} | Should -Not -BeNullOrEmpty
 		}
 		It 'finds the locally installed package just installed' {
-			Get-Package -Provider $WinGet -Name $package | Where-Object {$_.Name -contains $package} | Should -Not -BeNullOrEmpty
+			Get-Package -Provider $WinGet -Name $package -RequiredVersion $version | Where-Object {$_.Name -contains $package} | Should -Not -BeNullOrEmpty
 		}
 		It 'silently uninstalls the locally installed package just installed' {
-			Uninstall-Package -Provider $WinGet -Name $package | Where-Object {$_.Name -contains $package} | Should -Not -BeNullOrEmpty
+			Uninstall-Package -Provider $WinGet -Name $package -RequiredVersion $version | Where-Object {$_.Name -contains $package} | Should -Not -BeNullOrEmpty
 		}
 	}
 }
@@ -68,6 +73,7 @@ Describe "multi-source support" {
 		$altSourceName = 'AltWinGetSource'
 		$altSourceLocation = 'https://winget.azureedge.net/cache'
 		$package = 'CPUID.HWMonitor'
+		$version = '1.44'
 
 		Unregister-PackageSource -Name $altSourceName -Provider $WinGet -ErrorAction SilentlyContinue
 	}
@@ -82,7 +88,7 @@ Describe "multi-source support" {
 		Register-PackageSource -Name $altSourceName -Provider $WinGet -Location $altSourceLocation | Where-Object {$_.Name -eq $altSourceName} | Should -Not -BeNullOrEmpty
 	}
 	It 'searches for and installs the latest version of a package from an alternate source' {
-		Find-Package -Provider $WinGet -Name $package -source $altSourceName | Install-Package -Force | Where-Object {$_.Name -contains $package} | Should -Not -BeNullOrEmpty
+		Find-Package -Provider $WinGet -Name $package -Source $altSourceName -RequiredVersion $version | Install-Package -Force | Where-Object {$_.Name -contains $package} | Should -Not -BeNullOrEmpty
 	}
 	It 'unregisters an alternative package source' {
 		Unregister-PackageSource -Name $altSourceName -Provider $WinGet