+++ src/private/ConvertTo-SoftwareIdentity.ps1
@@ -24,7 +24,6 @@ function ConvertTo-SoftwareIdentity {
 			)
 			if ($packageSource) {
 				Write-Debug "Package identified: $($package.ID), $($package.version), $($packageSource)"
-				$metadata = Cobalt\Get-WinGetPackageInfo -ID $package.ID -Version $package.Version -Source $packageSource
 				$swid = @{
 					FastPackageReference = $package.ID+"#"+ $package.version+"#"+$packageSource
 					Name = $package.ID
@@ -32,9 +31,14 @@ function ConvertTo-SoftwareIdentity {
 					versionScheme = "MultiPartNumeric"
 					FromTrustedSource = $true
 					Source = $packageSource
-					Summary = $metadata.Description
-					FullPath = $metadata.'Download URL'
 				}
+
+				if ($request.Options.ContainsKey($script:Detailed)) {
+					$metadata = Cobalt\Get-WinGetPackageInfo -ID $package.ID -Version $package.Version -Source $packageSource
+					$swid.Summary = $metadata.Description
+					$swid.FullPath = $metadata.'Download URL'
+				}
+
 				New-SoftwareIdentity @swid
 			}
 		}