+++ src/public/BaseFunctions.ps1
@@ -22,9 +22,11 @@ function Get-DynamicOptions {
 	switch ($category) {
 		Package {
 			Write-Output -InputObject (New-DynamicOption -Category $category -Name $script:AcceptLicense -ExpectedType Switch -IsRequired $false)
+			Write-Output -InputObject (New-DynamicOption -Category $category -Name $script:Detailed -ExpectedType Switch -IsRequired $false)
 		}
 		Install {
 			Write-Output -InputObject (New-DynamicOption -Category $category -Name $script:AcceptLicense -ExpectedType Switch -IsRequired $false)
+			Write-Output -InputObject (New-DynamicOption -Category $category -Name $script:Detailed -ExpectedType Switch -IsRequired $false)
 		}
 	}
 }