+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ### Added
 * Include package summary and download URL in software identities
 * Support for the Find-Package `AllVersions` parameter
+* Support for the `RequiredVersion` parameter across multiple cmdlets
 
 ## 0.0.6 - 2022-02-06 - Force dependency checks
 ### Fixed