+++ CHANGELOG.md
@@ -7,9 +7,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## 0.0.7 - 2022-03-12 - Additional Package Details
 ### Added
-* Include package summary and download URL in software identities
+* Include package summary and download URL in package data
 * Support for the Find-Package `AllVersions` parameter
-* Support for the `RequiredVersion` parameter across multiple cmdlets
+* Support for the `RequiredVersion`, `MinimumVersion`, and `MaximumVersion` parameters across multiple cmdlets
 
 ## 0.0.6 - 2022-02-06 - Force dependency checks
 ### Fixed